requirements.txt
asgiref==3.7.2
Django==4.2.3
django-request-logging==0.7.5
djangorestframework==3.14.0
install==1.3.5
Pillow==10.0.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
